RDKTV-13125: WPEFramework timeout during shutdown in Maintenance cleanup
Reason for change: Graceful exit is not happening with maintenace manager
while stopping wpeframework service when maintenance is in progress. It got
stuck with deInitialize method as it was locked until a conditional variable
is notified. Hence notified the waiting tasks.
Test Procedure: Maintenance manager should exit gracefully when it is in progress
Risks: Medium

Signed-off-by: Pavithra <Pavithra_Pandian@comcast.com>
Change-Id: I7129fa766e2a3c097868faa0ca39d1aebe73c643