RDKTV-15888,RDKTV-15391,RDKTV-15673: Unregister when Socket Closed
Reason for change: When the socket is closed, the object pointer is not
removed from the ResourceMonitor. This is leading to dereference the
pointer that got destructed but not removed from ResourceMonitor when
there is race between closing the socket and some pending event to be
performed on the socket.

Test Procedure: As per RDKTV-15888
Risks: Medium
Signed-off-by: Zameerun Rasheed M S <zmamoo711@cable.comcast.com>

(cherry picked from commit 00046fd9f99bf06807e2c1c6da58ecda631c62af)
Change-Id: I9b08918406251f8803bbea6e1f1b72cdc453b7c9