DELIA-58794: Recover LC103s in OTA stuck state
Reason for change:  When an LC103 running fw older than 5103.2.4 is in
an OTA stuck state due to a previous OTA attempt being canceled mid-way,
it can be recovered if a button is pressed during another OTA attempt.
So, this change identifies when a remote is in that state, and triggers
another OTA on a keypress.  This will increase the likelyhood that a
user will press a button during the 30 second timeout window of an OTA
Test Procedure:     see ticket
Risks:              Low

Signed-off-by: Gene Gallagher <Eugene_Gallagher@comcast.com>

LLAMA-8671:i Prevent LC103 from sleeping during OTA

Reason for change:  In LC103s running fw < 5103.3.2, there's a
possibility they can go to sleep in the time between it receives 100% of
the OTA packets and when it sets the new image as acive in the
bootloader.  If this happens, the remote is essentially bricked until
the batteries are pulled for 5 minutes.

To prevent the remote from sleeping, the BLE connection parameters are
changed to their minimum values during an OTA. When the remote reboots,
it sets the connection parameters back to their original values.
Test Procedure:     see ticket
Risks:              Low

Change-Id: I88118882928698aaae489e1f8f55e00e0d847348
Signed-off-by: Gene Gallagher <Eugene_Gallagher@comcast.com>

LLAMA-5743: Enable standby mode C for LC103s on >= 2.6

Reason for change:  When reading all attributes from a remote during
initialization, if the remote is disconnected then a bunch of attributes
will be received as empty or invalid.  In this case, keep the values we
have previously stored in the ctrlm db
Test Procedure:     see ticket
Risks:              Low

Change-Id: Ie19113fe864e0ceddf0584107a7984943ddc5526
Signed-off-by: Gene Gallagher <Eugene_Gallagher@comcast.com>

DELIA-58494: add BLE pairing support for XR103 remote

Reason for change:  LC104 Flex 2.0 remote will be called XR103
Test Procedure:     XR103 pairs, xconf metadata is properly exported
Risks:              Low

Change-Id: Ia75064f43220fbd4b3aaefa16f18ecc6e4b6be3c
Signed-off-by: Gene Gallagher <Eugene_Gallagher@comcast.com>